Move scheduler/action fixtures to conftest.py
This is in preparation for testing other parts of the system.
